+++ src/solverpy/solver/atp/vampire.py
@@ -1,8 +1,10 @@
 from typing import TYPE_CHECKING
 import re
 
-from ..tptpsolver import TptpSolver
+from ..shellsolver import ShellSolver
 from ...tools import patterns, human
+from ..plugins.status.tptp import Tptp
+from ..plugins.shell.time import Time
 
 if TYPE_CHECKING:
    from ..plugins.plugin import Plugin
@@ -30,7 +32,7 @@ V_TABLE = {
 }
 
 
-class Vampire(TptpSolver):
+class Vampire(ShellSolver):
 
    def __init__(
       self,
@@ -41,7 +43,11 @@ class Vampire(TptpSolver):
       plugins: list["Plugin"] = [],
    ):
       cmd = f"{binary} {static}"
-      TptpSolver.__init__(
+      plugins = plugins + [
+         Time(),
+         Tptp(complete=complete),
+      ]
+      ShellSolver.__init__(
          self,
          cmd,
          limit,