+++ src/solverpy/solver/plugins/status/smt.py
@@ -6,14 +6,44 @@ from ....tools import patterns
 
 if TYPE_CHECKING:
    from ....tools.typing import Result
+   from ...solverpy import SolverPy
 
 SMT_STATUS = re.compile(r"^(sat|unsat|unknown|timeout)$", re.MULTILINE)
 
+SMT_OK = frozenset([
+   'sat',
+   'unsat',
+])
+
+SMT_FAILED = frozenset([
+   'unknown',
+])
+
+SMT_TIMEOUT = frozenset([
+   'timeout',
+   'memout',
+   'TIMEOUT',  # simulated timeout
+])
+
+SMT_INCOMPLETE = frozenset([
+   "sat",
+])
+
+SMT_ALL = SMT_OK | SMT_FAILED | SMT_TIMEOUT
+
+SMT_INC_OK = SMT_OK - SMT_INCOMPLETE
 
 class Smt(Decorator):
 
-   def __init__(self, **kwargs):
+   def __init__(self, complete=True, **kwargs):
       Decorator.__init__(self, **kwargs)
+      self._complete = complete
+
+   def register(self, solver: "SolverPy"):
+      super().register(solver)
+      solver._success |= SMT_OK if self._complete else SMT_INC_OK
+      solver._timeouts |= SMT_TIMEOUT
+      solver._statuses |= SMT_ALL
 
    def decorate(
       self,